+++ src/main/java/net/vexelon/bgrates/Defs.java
@@ -41,7 +41,7 @@ public class Defs {
 	// Name of DB
 	public static final String DATABASE_NAME = "currencies.db";
 	// Version of DB
-	public static final int DATABASE_VERSION = 1;
+	public static final int DATABASE_VERSION = 2;
 	// table name
 	public static final String TABLE_CURRENCY = "currencies";
 	public static final String TABLE_CURRENCY_DATE = "currenciesdate";